Add static pages for a typical website (#4)
Adds basic pages for a static website. The primary goal is to get this
deployed and working as a test site, including all images, SEO
characteristics, caching, PageSpeed scores > 98, etc.

Once that is working, we'll work backwards to making the site dynamic,
which will allow for other sites to be rendered in exactly the same way.

Future PRs will get it running in production. Then we can start
refactoring to use the normal assets pipeline and RESTful routes, etc.