Port vanilla JS to StimulusJS (#10)
Moves the JS from the `public/assets/js` folder to the
`app/javascript/controllers` folder.

The dark mode toggle is working as expected, although that feature might
be removed.

The navigation JS is commmented for now because there are no dropdown
items in the navigation, so it's not needed at the moment. When I
refactor the navigation, I'll uncomment it and make sure it works.