Remove old assets (#11)
We no longer need the `temp_assets_directory`, so this removes the
workarounds put in place to make initial development easier.

I also removed the favicon setup entirely, because I'm not sure how
necessary any of that will be. In any case, I'd want to replace the
provided favicons with something more relevant, so there's no harm in
just sidestepping the whole issue right now.